game is no longer limited to max 255 items. Really xD

+++ src/dfdata.pas
@@ -312,7 +312,7 @@ type TItemProperties = record
   Durability    : Word;
   MaxDurability : Word;
 
-  AmmoID        : Byte;
+  AmmoID        : Integer;
   Ammo          : Word;
   AmmoMax       : Word;
   Acc           : Integer;
